requirements-dev.txt
-r requirements.txt
pre-commit==4.1.0
ruff==0.9.6
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
mypy==1.14.1
django-stubs==5.1.2
